recursive_diagrams/pyproject.toml
[project]
name = "recursive-diagrams"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "kagglehub>=0.4.3",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "ucimlrepo>=0.0.7",
    "wfdb>=4.3.0",
    "yfinance>=1.1.0",
]
